+++ src/engine.js
@@ -4537,9 +4537,7 @@ var XPathJS = (function(){
 				{
 					// Using ParseInt, creates a problem for very small or large numbers that are displayed in scientific
 					// notation. E.g. parseInt(1/47999799999, 10) is 2 instead of 0 (2.08e-11)
-					// Therefore this function is essentially an alias for floor(). It's argument is a string though, 
-					// but I don't think it behaves different from floor in any way.
-					return new NumberType( Math.floor( str.toNumber() ) );
+					return new NumberType( ( str.toNumber() >= 0 ) ? Math.floor( str.toNumber() ) : -Math.floor( Math.abs( str.toNumber() ) ) );
 				}, 
 				args: [
 					{t: 'string'}